#  Return request data to the pool in Ingester.Push()