--- a/pkg/ingester/ingester.go
+++ b/pkg/ingester/ingester.go
@@ -279,6 +279,7 @@ func (i *Ingester) Push(ctx old_ctx.Context, req *client.WriteRequest) (*client.
 			return nil, err
 		}
 	}
+	client.ReuseSlice(req.Timeseries)
 
 	return &client.WriteResponse{}, lastPartialErr
 }