After #1526 the unmarshall code is trying to get memory out of the pool, but in ingesters it is never returned so the `pool.Get()` call just slows things down.

After making this change it still wasn't that great, because the pool was boxing `TimeSeries` objects.  So change those to `*TimeSeries`.

before:
```
BenchmarkIngesterPush-2   	     100	  23525939 ns/op	 3881766 B/op	   42317 allocs/op
```

after:
```
BenchmarkIngesterPush-2   	     100	  22937648 ns/op	 3273382 B/op	   32482 allocs/op
```

(note out of the 3-4MB per iteration, 1MB is overhead from the `FingerprintLocker`)